Add support for Hue buttons
These can be used to manually enable ventilation of a room for a fixed
amount of time. An example of where this is useful is a kitchen
extractor: detecting when to enable this automatically is more or less
impossible, so instead we can use a button to enable it for a certain
amount of time.

Both short and long presses are supported, though a room can only react
to a single type of press. In addition, as we poll the state rapidly
pressing the button different ways results in us only observing the most
recent state.