Basic ventilation when humidity increases
Humid air can seemingly build up in the ventilation duct of a bathroom.
If enough such air builds up, we may end up ventilating the room using
the "Humid" state, leading to an unnecessary amount of noise being
produced.

Instead, when the humidity is between the "low" and "high" thresholds,
we apply regular ventilation of the room. This should keep the humidity
under control, without the noise.