Remove support for Hue buttons
I got a button to connect to a bridge once, but it stopped working the
next day. Since then I haven't been able to connect it, a second button,
or a dimmer switch. Having tried everything, I'm throwing in the towel
and will look into other options for enabling the extractor fan on
demand.