Add support for per-room flow correction
When multiple rooms are configured at the same flow rate (e.g. 40 m3/h),
they may have actual flow rates that slightly differ. For example, if
for one room the air has to pass through more corners in the ducts, this
may result in a slight loss of air flow.

To compensate for this a "flow correction" can be configured. This value
is multiplied with the configured air flow to derive the air flow that
should be applied instead. This ensures that the same configured air
flow results in the same end air flow across all rooms.