Reduce the maximum exhaust value to 4000
When measuring the maximum air flow through my exhausts, I observed no
difference in speed between a setting of 4000 and 5000. My DemandFlow
system also appears to be configured to use 4000 as the maximum, instead
of 5000. Finally, when using 5000 as the setting the servos sometimes
make a distinct "tick tick tick" noise, likely due to the rubber covers
getting stretched too much.

To fix all this we now use 4000 as the maximum valve setting.