Remove ventilation above humidity low thresholds
The code was supposed to only start ventilating if the humidity levels
were above the low threshold long enough. Due to the way the code was
written it would instead immediately start ventilating the room the
moment the threshold was crossed.

In practise this doesn't seem to do anything other than sporadically
increase ventilation for 10-15 minutes. Instead we can just rely on
minimum ventilation, so let's get rid of this code entirely.