Wait for CO2 levels to stabilise before sampling
When sampling rooms it can take 2-3 minutes before the CO2 levels in the
plenum are the same as in the room that's sampled. If we record samples
during this time, the end CO2 value may be biased towards the CO2 levels
of unrelated rooms.

To solve this we split the sampling time into two phases: first we wait
X seconds without recording samples, then we sample for Y seconds while
recording the samples. Both these times are configured separately.