Ventilate at lower speeds when CO2 levels are low
When the CO2 levels in a room are below the target CO2 level, we now
ventilate at half the speed. This reduces noise when many rooms are
ventilated but the CO2 levels are low, such as when the doors are all
opened. The moment we reach or cross the target CO2 level, ventilation
runs at the usual speed.