Tweak the speed in response to CO2 levels
Rather than running at the configured speed when reaching the target CO2
concentration, we run at half speed, only switching to the configured
speed when the CO2 levels are _above_ the target threshold.

In other words: if the target is 700 ppm, then we run at half the speed
for CO2 levels <= 700 ppm. If the CO2 concentration is 750 ppm we run at
the configured speed, and for higher levels we apply the CO2 increase.

The goal is to reduce the necessary air flow for larger rooms, as CO2
levels are less likely to build up much/quickly in such rooms, and thus
we can permit lower ventilation speeds.