Enable the night schedule based on CO2 levels
It's not unlikely for the night schedule to active before a bedroom is
in use, such as when you're out or simply going to bed late. This
changes the schedule handling such that the night schedule is only
active as long as the CO2 levels are at or above the target CO2 levels.