Don't enable the night schedule based on CO2
This creates a circular dependency: rooms of which the state is Closed
are only sampled every 3 sampling intervals (so every 3 hours in case of
a 1 hour interval), meaning that rooms with the Night schedule may need
to wait up to 3 such intervals to receive ventilation.

Thinking about this, this isn't that necessary either: if the CO2 levels
are way below the target level (which would be the case until a person
sleeps in the room for a bit), ventilation levels are already reduced,
and thus noise is also low.