Adjust air flow to achieve a target CO2 level
Instead of increasing and decreasing the ventilation speed as CO2 levels
change, we dynamically adjust the speed until we reach a speed capable
of keeping CO2 levels at a target value.

This is still experimental and may be removed if it turns out to not
work as well as hoped.