Simplify scaling CO2 ventilation speeds
The old algorithm was interesting on paper, but in practise took too
long to scale up the ventilation speed when needed. I also found that in
most cases finding the right speed is quite easy: when doors are open or
rooms are large and only a single person is in the room, 30 m3/h is more
than enough. When two or more people are in the room or the doors are
closed, it seems 60-70 m3/h is needed instead (based on the room size).

This new algorithm is a bit simpler and should better cater towards
this. Below or at the target value we run at the configured speed. Above
the target value we simply double the ventilation speed, tripling it if
we are instead above a "high" threshold (which defaults to 1000 ppm).
Similar to the old algorithm we then maintain said speed until we're
either below the "high" threshold, or below the target, at which point
we revert to whatever speed is needed for that level.

In practise this should result in the system doubling the ventilation
speed the moment CO2 levels go above the target threshold, then maintain
that until the person leaves, opens the door, etc.