Ventilate closed rooms when CO2 levels are high
This is useful for e.g. a living room with an open kitchen. In this case
the kitchen may report the same high CO2 levels as the living room.
Enabling the ventilation in this case can help reduce CO2 levels,
without too much noise being produced by a single exhaust.