Separate minimum and default air flows
This better captures what the two mean: minimum flow is a hard minimum,
and default flow is used when no active ventilation is needed, and is
still subject to the minimum flow.

The use case is a bathroom that you always want to run at e.g. 10 m3/h,
regardless of what kind of ventilation is necessary for other rooms.