Simplify CO2 speed calculations
Gradually increasing the speed per a certain number of CO2 buckets
doesn't seem to do much, as most of the time we end up with the same 2-3
speeds (basically a low, medium and high speed). In this new setup we
use fixed speeds for CO2 levels, which thus far is working much better.