Remove outlier removal
When entering a small room and closing the door, CO2 levels rise quickly
over the course of 10-15 minutes. By removing outliers, it's possible
active ventilation is delayed for too long. At least, that's what I
suspect, so this commit is an attempt at trying to determine if that's
true or not over the course of a few days.