Handle humidity increases below the threshold
When the indoor air is dry (e.g. during the summer) and one takes a
shower, the humidity might increase but not enough to cross the absolute
threshold.

To solve this, the humidity input maintains a ring buffer of humidity
samples. When the most recent sample is greater than a certain
percentage (15% by default) compared to 5 minutes ago, ventilation is
enabled regardless of the absolute humidity. So if the humidity
increases from 55% to 70%, but the threshold is 80%, humidity is still
enabled.

This increase acts as an OR with the absolute threshold, ensuring that
if humidity slowly increases we still enable ventilation when necessary.